Rename Controller to Repository
