+++ src/main/kotlin/com/tedm/data/repository/user/UserRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.tedm.data.controller.user
+package com.tedm.data.repository.user
 
 import com.tedm.data.models.User
 import org.litote.kmongo.coroutine.CoroutineDatabase
