+++ src/main/kotlin/com/tedm/plugins/Routing.kt
@@ -1,11 +1,14 @@
 package com.tedm.plugins
 
+import com.tedm.data.repository.user.UserRepository
 import com.tedm.routes.createUserRoute
 import io.ktor.routing.*
 import io.ktor.application.*
+import org.koin.ktor.ext.inject
 
 fun Application.configureRouting() {
+    val userRepository: UserRepository by inject()
     routing {
-        createUserRoute()
+        createUserRoute(userRepository)
     }
 }
